+++ qgreenland/config/layers/Reference/Borders/political_boundaries.py
@@ -10,20 +10,27 @@ nunagis_municipalities_population = Layer(
     title="Greenland municipalities and population",
     description=(
         """Polygons representing municipalities of Greenland and associated
-        population numbers for 2019 and 2025."""
+        population numbers from 1976-2024.
+
+        Note: to use this layer, enable the temporal controller and set a
+        specific year to see that year's population statistics. Without the
+        temporal controller enabled, the earliest population statistics will be
+        shown (for 1976).
+        """
     ),
     tags=[],
-    style="nunagis_municipalities_population",
+    style="municipalities_pop_timeseries",
     inputs=[
         # This input provides a multipolygon of municipalities and population numbers for 2019
         LayerInput(
             dataset=political_boundaries.nunagis_pop2019_municipalities,
             asset=political_boundaries.nunagis_pop2019_municipalities.assets["only"],
         ),
-        # This input provides updated population statistics for 2025 (Jan 1, 2026).
+        # This input provides updated population statistics for 1977-2025 (each
+        # stat valid for Jan 1 of the associated year).
         LayerInput(
             dataset=statbank.statbank,
-            asset=statbank.statbank.assets["municipalities_2025_population"],
+            asset=statbank.statbank.assets["municipalities_population"],
         ),
     ],
     steps=[
@@ -41,7 +48,7 @@ nunagis_municipalities_population = Layer(
                 "{output_dir}/merged.gpkg",
                 "{input_dir}/BEXSTB.csv",
                 "-nln",
-                "pop2025",
+                "pop",
             ],
         ),
         CommandStep(
@@ -54,16 +61,96 @@ nunagis_municipalities_population = Layer(
                 "-dialect",
                 "sqlite",
                 "-sql",
+                # The population statistics are valid for Jan 1st of the
+                # indicated year. For use with the temporal controller, we set
+                # start/end dates to reflect the prior year.
                 """\"SELECT
                     municipalities.geom,
                     municipalities.pop_municipality_2019_municip as municipality,
-                    municipalities.pop_municipality_2019_populatio as \\"Population 2019\\",
-                    pop2025.\\"Population January 1st\\" as \\"Population 2025\\"
+                    pop.\\"Population January 1st\\" as population,
+                    DATE(CAST(CAST(pop.time AS int) - 1 AS STR) || '-01-01') as valid_start_date_str,
+                    DATE(CAST(CAST(pop.time AS int) - 1 AS STR) || '-12-31') as valid_end_date_str
                     FROM municipalities
-                    JOIN pop2025 ON pop2025.municipality
-                    LIKE '%' || municipalities.pop_municipality_2019_municip || '%'\"""",
+                    RIGHT JOIN pop ON pop.municipality
+                    = municipalities.pop_municipality_2019_municip\"""",
                 "-nln",
                 "municipalities_and_pop",
+            ],
+        ),
+        # This step creates a new `valid_date` column with the `Date` field
+        # type. This has to be done separately because sqlite will cast `DATE()`
+        # (selected above) to text type.
+        # TODO: this should be a one-liner with `gdal vector set-field-type`,
+        # but that requires gdal >=3.12, and we are currently stuck at gdal 3.10
+        # because of conflicts with other dependencies (I think pydantic needs
+        # to be updated to v2!). It is possible to update gdal to 3.12 in the
+        # command environment, which I think would be OK but would lead to a
+        # disconnect between the gdal in the main env and the command env. From
+        # what I can tell, gdal is only used for generating statistics in the
+        # `main` env, which could be accomplished via a subprocess...
+        CommandStep(
+            id="update_date_field",
+            args=[
+                # Copy the input data to the output location. The subsequent
+                # commands will update the data in-place
+                "cp {input_dir}/joined.gpkg {output_dir}/final.gpkg",
+                "&&",
+                # update the municipalities_and_pop table to include a start
+                # date column with `DATE` type.
+                "ogr2ogr",
+                "-update",
+                "{output_dir}/final.gpkg",
+                "{output_dir}/final.gpkg",
+                "-dialect",
+                "sqlite",
+                "-sql",
+                """\"ALTER TABLE municipalities_and_pop
+                    ADD COLUMN start_date DATE\"""",
+                "&&",
+                # update the municipalities_and_pop table to include a end date
+                # column with `DATE` type.
+                "ogr2ogr",
+                "-update",
+                "{output_dir}/final.gpkg",
+                "{output_dir}/final.gpkg",
+                "-dialect",
+                "sqlite",
+                "-sql",
+                """\"ALTER TABLE municipalities_and_pop
+                    ADD COLUMN end_date DATE\"""",
+                "&&",
+                # Set the column w/ values from the str field.
+                "ogr2ogr",
+                "-update",
+                "{output_dir}/final.gpkg",
+                "{output_dir}/final.gpkg",
+                "-dialect",
+                "sqlite",
+                "-sql",
+                """\"UPDATE municipalities_and_pop
+                    SET start_date = valid_start_date_str, end_date = valid_end_date_str\"""",
+                "&&",
+                # drop the str date fields, which is now unnecessary
+                "ogr2ogr",
+                "-update",
+                "{output_dir}/final.gpkg",
+                "{output_dir}/final.gpkg",
+                "-dialect",
+                "sqlite",
+                "-sql",
+                """\"ALTER TABLE municipalities_and_pop
+                     DROP COLUMN valid_start_date_str\"""",
+                "&&",
+                # drop the str date fields, which is now unnecessary
+                "ogr2ogr",
+                "-update",
+                "{output_dir}/final.gpkg",
+                "{output_dir}/final.gpkg",
+                "-dialect",
+                "sqlite",
+                "-sql",
+                """\"ALTER TABLE municipalities_and_pop
+                     DROP COLUMN valid_end_Date_str\"""",
             ],
         ),
     ],