+++ CHANGELOG.md
@@ -1,11 +1,15 @@
-# v4.0.0alpha2 (TBD)
+# v4.0.0alpha2 (2025-03-13)
 
 - Minimum supported version of QGIS is now `v3.44`.
 - Update layers:
+  - "Reference/Borders/Greenland municipalities and population". Updated to
+    include population statistics from 1977-2025. This layer is now enabled for
+    use with the temporal controller, which should be enabled for proper
+    visualization!
   - "Glaciology/Gravimetric mass balance/Mass balance trend" layers have updated
     dataset metadata (title and citation URL).
   - "Biology/Birds/Murre colonies": data and metadata updated to Arctic
-    Biodiversity Assessment 2013 publication from CAFF .
+    Biodiversity Assessment 2013 publication from CAFF.
   - Replaced the "Biology/Birds/Goose protected areas" layer from NunaGIS with
     two updated layers from NatureMap: "Barnacle goose colonies" (Point) and
     "Goose moulting and breeding areas" (Polygon).
@@ -34,6 +38,7 @@
     - "Atlantic salmon in Kapisillit" from NatureMap
   - "Biology/Vegetation":
     - "Important areas red listed plant taxa" from NatureMap
+
 
 # v4.0.0alpha1 (2026-03-04)